refactor(epicenter): delete unused sync-test-utils.ts, inline utilities
